chore: reorganizing code for better maintainability
