+++ lib/index.js
@@ -1,9 +1,9 @@
-const { currentTemperature } = require('./currentTemperature');
-const { findCoordinates } = require('./findCoordinates');
-const { temperatureSelector } = require('./functionSelector');
+const { currentTemperature, hourlyTemperature } = require('./temperature/index');
+const { findCoordinates, temperatureSelector } = require('./utils/index');
 
 module.exports = {
     currentTemperature,
+    hourlyTemperature,
     findCoordinates,
     temperatureSelector,
 }
