+++ lib/temperature/currentTemperature.js
@@ -1,5 +1,5 @@
 const axios = require('axios');
-const { findCoordinates } = require('./findCoordinates');
+const { findCoordinates } = require('../utils/findCoordinates');
 
 async function currentTemperature(city) {
     try {
