+++ lib/temperature/index.js
@@ -1,0 +1,7 @@
+const { currentTemperature } = require('./currentTemperature');
+const { hourlyTemperature } = require('./hourlyTemperature');
+
+module.exports = {
+    currentTemperature,
+    hourlyTemperature
+}
