+++ lib/utils/findCoordinates.js
