+++ lib/utils/index.js
@@ -1,0 +1,7 @@
+const { findCoordinates } = require('./findCoordinates');
+const { temperatureSelector } = require('./selector');
+
+module.exports = {
+    findCoordinates,
+    temperatureSelector
+}
