+++ lib/utils/selector.js
@@ -1,5 +1,4 @@
-const { currentTemperature } = require('./currentTemperature');
-const { hourlyTemperature } = require('./hourlyTemperature');
+const { currentTemperature, hourlyTemperature } = require('../temperature/index');
 
 async function temperatureSelector (args, options ) { 
     if (options.current) {
